task 5 done

+++ src/main/java/com/jpmc/midascore/entity/UserRecord.java
@@ -13,19 +13,19 @@ public class UserRecord {
     private String name;
 
     @Column(nullable = false)
-    private float balance;
+    private double balance;
 
     protected UserRecord() {
     }
 
-    public UserRecord(String name, float balance) {
+    public UserRecord(String name, double balance) {
         this.name = name;
         this.balance = balance;
     }
 
     @Override
     public String toString() {
-        return String.format("User[id=%d, name='%s', balance='%f'", id, name, balance);
+        return String.format("User[id=%d, name='%s', balance='%.2f'", id, name, balance);
     }
 
     public Long getId() {
@@ -36,11 +36,11 @@ public class UserRecord {
         return name;
     }
 
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(float balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 }
